==> code/final/00_required_packages.R <==
install.packages("tidyverse")
install.packages("arrow")
install.packages("data.table")
install.packages("janitor")
install.packages("here")
install.packages("summarytools")
install.packages("patchwork")
install.packages("trelliscopejs")
install.packages("plotly")
install.packages("flextable")
install.packages("rvg")
install.packages("officer")
install.packages("DT")


# Important links ---------------------------------------------------------

## Original NGSIM data: https://datahub.transportation.gov/stories/s/Next-Generation-Simulation-NGSIM-Open-Data/i5zb-xe34/
## Interstate 80 Freeway factsheet: https://www.fhwa.dot.gov/publications/research/operations/06137/index.cfm
## Random forest car-following model paper: https://www.scirp.org/journal/paperinformation.aspx?paperid=110945
